+++ notes/logic.cpp
@@ -516,7 +516,7 @@ void Notes_Logic::emailUsers (int identifier, const string& label, string bible,
   int timestamp = filter::date::seconds_since_epoch ();
   if (postpone) {
     int localseconds = filter::date::local_seconds (timestamp);
-    float localhour = filter::date::numerical_hour (localseconds) + (float) filter::date::numerical_minute (localseconds) / 60;
+    float localhour = static_cast<float>(filter::date::numerical_hour (localseconds)) + (float) filter::date::numerical_minute (localseconds) / 60;
     if (localhour < 21) {
       float difference = 21 - localhour;
       timestamp += static_cast<int>(3600 * difference) - 10;
